Left click pauses, check new unit in pause.
Left click should pause, but chromebook has crappy left click so
it's hard for me to be sure.  When paused, you should be able to
now select a new unit and issue it an order.  This should allow
issuing of multiple orders...though as soon as I click where to
move, it will unpause.  Consider changing pause to spacebar, or p
or something.  clicking a unit will still pause the game, but only
clicking a 'pause key' will unpause it.