Added basic combat window.  And fixed combat bug.
Super basic, like literally a white window pops up in battle.  No
units or anything yet.

Combat sometimes seemed to last forver.  It appears that I wasn't
ever actually setting the number of attacks each unit had available
to use.  Now that's set at the beginning of battle.  So it could
change with a status effect of some kind.  That'd require passing
in some game state, like a 'fight state' class or something that
had both of the units or something.