Set attacker/defender and fix no defender bug.
The heroes who attack and defend are now set.  In the future,
return a list of defenders in case we attack all enemies.  At least
, we can apply the attack to each enemies.

Bug was related to trying to set defender when there was no one
left to defend.  Changed so we only do that and update the window
when there's an actual defender.