Moved fighting before moving.
This fixes the bug that had highlighted fighting units ghosted in
two places.  Because we fight first, we're in the same position
for the highlight as we were when displayed at the end of last
time step.

Still a minor problem of units that start within fighting distance,
the screen is all black except for them, but this is kind of cool.
It's a fight in media res.