Move around hero avatars in fight
Also make the leader a random unit instead of front...
This could induce some bugs, don't know yet.

Ideally, put the avatar on top of the polygons, make the leader
larger (or give the leader special units that the others can't be).
So maybe I need to add one more piece of art, a 'grunt'.  Also
improve the leader's stats?